+++ front/src/component/News/styles/Dot.module.scss
@@ -1,0 +1,36 @@
+@use "../../../styles/mixin";
+@use "../../../styles/color";
+@use "../../../styles/mediaQuery";
+@use "../../../styles/duration";
+
+.Dot {
+    @include mixin.clickable;
+
+    width: 1.25rem;
+    height: 1.25rem;
+    opacity: 0.65;
+    border-width: 1px;
+    border-radius: 999999px;
+    background-color: color.$bg-third;
+    border-color: color.$prime-darker;
+    transition: duration.$base background-color;
+
+    @media (width > mediaQuery.$md) {
+        width: 1rem;
+        height: 1rem;
+    }
+
+    &:hover {
+        background-color: color.$bg-second;
+    }
+
+    &.current {
+        opacity: 1;
+        border-color: color.$prime-darker;
+        background-color: color.$prime;
+
+        &:hover {
+            background-color: color.$prime-darker;
+        }
+    }
+}
