+++ front/src/component/News/styles/Dots.module.scss
@@ -1,0 +1,4 @@
+.Dots {
+    display: flex;
+    gap: 0.5rem;
+}
