+++ front/src/component/News/styles/News.module.scss
@@ -1,0 +1,8 @@
+.News {
+    width: 100%;
+    height: 28rem;
+    display: flex;
+    gap: 0.5rem;
+    place-items: center;
+    place-items: center;
+}
